Add new maplayer component HexGrid
