+++ frontapp/src/components/Map/index.tsx
@@ -1,14 +1,14 @@
 'use client';
 
 import { MapContainer, TileLayer, Marker, Popup } from 'react-leaflet';
-import { GeoJSON } from 'react-leaflet/GeoJSON';
+import HexGrid from '../MapLayers/HexGrid';
 import { LatLngExpression, LatLngTuple } from 'leaflet';
+import fortalezaHexagons from '../../../public/fortaleza-hexgrid.json';
 
 import 'leaflet/dist/leaflet.css';
 import 'leaflet-defaulticon-compatibility/dist/leaflet-defaulticon-compatibility.css';
 import 'leaflet-defaulticon-compatibility';
 import { useRef } from 'react';
-import hexagons from '../../../fortaleza-hexgrid.json';
 
 interface MapProps {
   posix: LatLngExpression | LatLngTuple;
@@ -32,20 +32,6 @@ const data: GeoJSON.Feature = {
   },
 };
 
-const datas: GeoJSON.FeatureCollection = hexagons as GeoJSON.FeatureCollection;
-
-const onEachFeature = (feature: any, layer: any) => {
-  let popupContent = `District Code: ${feature.properties.name} <br> District:`;
-  if (feature.properties && feature.properties.popupContent) {
-    popupContent += feature.properties.popupContent;
-  }
-  layer.on({
-    click: () => {
-      console.log('Feature clicked', feature.geometry.coordinates);
-    },
-  });
-};
-
 const Map = (Map: MapProps) => {
   const geoJsonLayerRef = useRef(data);
   const { zoom = defaults.zoom, posix } = Map;
@@ -57,11 +43,8 @@ const Map = (Map: MapProps) => {
       scrollWheelZoom={true}
       style={{ height: '100vh', width: 'auto' }}
     >
-      <GeoJSON
-        key={Math.random() + 2}
-        data={datas}
-        onEachFeature={onEachFeature}
-      />
+      <HexGrid data={fortalezaHexagons as GeoJSON.FeatureCollection} />
+
       <TileLayer
         attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
         url='https://{s}.tile.openstreetmap.org/{z}/{x}/{y}.png'
